myproject/requirements.txt
# Django and Core Dependencies
Django>=6.0.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
whitenoise>=6.6.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0

# AI/ML Dependencies
groq>=1.0.0
scikit-learn==1.6.1
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
xgboost>=2.0.0

# Celery for async tasks
celery>=5.3.0
redis>=4.5.0

# Utils
requests>=2.31.0
python-dateutil>=2.8.0
